requirements.txt
aiogram==2.22.2
python-dotenv==1.0.0